Deprecate tracksys in config.
To avoid confusion, configuration is only done for each stream. Flags for skipping interpolation in physio data and keeping time stamps for motion data are removed. By default time stamps (latencies) will be used for time synch for both motion and physio data. An exception is when physio data contains multiple streams and need to be concatenated. Stream separation for physio data is to be updated in the future.